# Implement ~70x Speed Increase for Encryption